fixed failing cypress and move necessary code utils
